server/requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "3.11"
appdirs==1.4.4 ; python_version >= "3.10" and python_version < "3.11"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.3 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
docker-pycreds==0.4.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.95.1 ; python_version >= "3.10" and python_version < "3.11"
gitdb==4.0.10 ; python_version >= "3.10" and python_version < "3.11"
gitpython==3.1.31 ; python_version >= "3.10" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
pathtools==0.1.2 ; python_version >= "3.10" and python_version < "3.11"
protobuf==4.22.3 ; python_version >= "3.10" and python_version < "3.11"
psutil==5.9.5 ; python_version >= "3.10" and python_version < "3.11"
pydantic==1.10.7 ; python_version >= "3.10" and python_version < "3.11"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "3.11"
requests==2.29.0 ; python_version >= "3.10" and python_version < "3.11"
sentry-sdk==1.21.0 ; python_version >= "3.10" and python_version < "3.11"
setproctitle==1.3.2 ; python_version >= "3.10" and python_version < "3.11"
setuptools==67.7.2 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
smmap==5.0.0 ; python_version >= "3.10" and python_version < "3.11"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.11"
starlette==0.26.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "3.11"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "3.11"
uvicorn==0.21.1 ; python_version >= "3.10" and python_version < "3.11"
wandb==0.15.0 ; python_version >= "3.10" and python_version < "3.11"
